text_to_image/requirements.txt
fastapi
uvicorn
boto3
botocore
pillow
pydantic
